+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Upgrade taskcat to 0.9.31
+* Fix cleanup task to include empty-and-delete-bucket.py
 
 # 2.0.0